pyproject.toml
[project]
name = "childrens-speech-recognition-runtime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "loguru>=0.7.3",
    "nemo-toolkit[all,asr]>=2.6.2",
    "omegaconf>=2.3.0",
    "pyannote-audio==3.3.2",
    "pytorch-lightning>=2.6.1",
    "sacrebleu>=2.6.0",
    "text-unidecode>=1.3",
    "wget>=3.2",
]
